M1-Challenge-Panlilio-Ian/Exercise 1/src/main/java/ConverterSwitch.java
public class ConverterSwitch implements Converter{
    @Override
    public String convertMonth(int monthNumber) {
        String monthString = "";
        switch (monthNumber) {
            case 1:
                monthString = "January";
                break;
            case 2:
                monthString = "February";
                break;
            case 3:
                monthString = "March";
                break;
            case 4:
                monthString = "April";
                break;
            case 5:
                monthString = "May";
                break;
            case 6:
                monthString = "June";
                break;
            case 7:
                monthString = "July";
                break;
            case 8:
                monthString = "August";
                break;
            case 9:
                monthString = "September";
                break;
            case 10:
                monthString = "October";
                break;
            case 11:
                monthString = "November";
                break;
            case 12:
                monthString = "December";
                break;
            default:
                monthString = "The number you entered does not correspond to a month of the year.";
        }
        return "ConverterSwitch: " + monthString;
    }
    @Override
    public String convertDay(int dayNumber) {
        String dayString = "";
        switch (dayNumber) {
            case 1:
                dayString = "Sunday";
                break;
            case 2:
                dayString = "Monday";
                break;
            case 3:
                dayString = "Tuesday";
                break;
            case 4:
                dayString = "Wednesday";
                break;
            case 5:
                dayString = "Thursday";
                break;
            case 6:
                dayString = "Friday";
                break;
            case 7:
                dayString = "Saturday";
                break;
            default:
                dayString = "The number you entered does not correspond to a day of the week.";
        }
        return "ConverterSwitch: " + dayString;
    }
}

